+++ src/map.ts
@@ -5,6 +5,7 @@ export { MapLineSeries, IMapLineSeriesSettings, IMapLineSeriesDataItem, IMapLine
 export { MapPointSeries, IMapPointSeriesSettings, IMapPointSeriesDataItem, IMapPointSeriesPrivate } from "./.internal/charts/map/MapPointSeries";
 export { MapPolygon, IMapPolygonSettings, IMapPolygonPrivate } from "./.internal/charts/map/MapPolygon";
 export { MapPolygonSeries, IMapPolygonSeriesSettings, IMapPolygonSeriesDataItem, IMapPolygonSeriesPrivate } from "./.internal/charts/map/MapPolygonSeries";
+export { ClusteredPointSeries, IClusteredPointSeriesSettings, IClusteredPointSeriesDataItem, IClusteredPointSeriesPrivate } from "./.internal/charts/map/ClusteredPointSeries";
 export { MapSeries, IMapSeriesSettings, IMapSeriesDataItem, IMapSeriesEvents, IMapSeriesPrivate } from "./.internal/charts/map/MapSeries";
 export { ZoomControl, IZoomControlSettings, IZoomControlPrivate } from "./.internal/charts/map/ZoomControl";
 export { getGeoRectangle, getGeoCircle, normalizeGeoPoint, getGeoCentroid, getGeoBounds, getGeoArea } from "./.internal/charts/map/MapUtils";